changed passport session strategy to store user info
