+++ passport/local-login.js
@@ -1,14 +1,33 @@
-
 const PassportLocalStrategy = require('passport-local').Strategy;
 const db = require('../models');
 const logger = require('winston');
-const { returnUserAndChannelInfo } = require('../helpers/authHelpers.js');
+
+function returnUserAndChannelInfo (userInstance) {
+  return new Promise((resolve, reject) => {
+    let userInfo = {};
+    userInfo['id'] = userInstance.id;
+    userInfo['userName'] = userInstance.userName;
+    userInstance
+        .getChannel()
+        .then(({channelName, channelClaimId}) => {
+          userInfo['channelName'] = channelName;
+          userInfo['channelClaimId'] = channelClaimId;
+          return db.Certificate.getShortChannelIdFromLongChannelId(channelClaimId, channelName);
+        })
+        .then(shortChannelId => {
+          userInfo['shortChannelId'] = shortChannelId;
+          resolve(userInfo);
+        })
+        .catch(error => {
+          reject(error);
+        });
+  });
+}
 
 module.exports = new PassportLocalStrategy(
   {
     usernameField: 'username',
     passwordField: 'password',
-    // session      : false,
   },
   (username, password, done) => {
     logger.debug('logging user in');
@@ -20,7 +39,7 @@ module.exports = new PassportLocalStrategy(
             // logger.debug('no user found');
             return done(null, false, {message: 'Incorrect username or password.'});
           }
-          return user.comparePassword(password, (passwordErr, isMatch) => {
+          user.comparePassword(password, (passwordErr, isMatch) => {
             if (passwordErr) {
               logger.error('passwordErr:', passwordErr);
               return done(null, false, {message: passwordErr});
